Deleted 'adding a new endpoint' page
This page has no content, was a placeholder, and contains outdated info already (eg Flask)
Suggest delete and reinstate only if there is a reason to bring it back. Not sure how often an external contributor is going to need to create a new API endpoint. Would be challenging to keep this page updated, when the FastAPI docs are probably going to be a more appropriate source of the information.